fix: move logic to separate file & test
